+++ src/serialization/__serialize_51.cpp
@@ -28,8 +28,8 @@ namespace fertilized {
     if (always_register ||
         serialization_library_version >= 0) {
       ar.template register_type<DifferenceSurfaceCalculator<
-              uint8_t,
-              int16_t,
+              float,
+              float,
               uint
 	  >>();
     }
@@ -44,18 +44,18 @@ namespace fertilized {
       const unsigned int &serialization_library_version);
 
     TemplateFuncExport DllExport std::string serialize(const DifferenceSurfaceCalculator<
-              uint8_t,
-              int16_t,
+              float,
+              float,
               uint
 	  > *, const bool &);
     TemplateFuncExport DllExport DifferenceSurfaceCalculator<
-              uint8_t,
-              int16_t,
+              float,
+              float,
               uint
 	  >* deserialize(std::stringstream &);
     TemplateFuncExport DllExport void deserialize(std::stringstream &, DifferenceSurfaceCalculator<
-              uint8_t,
-              int16_t,
+              float,
+              float,
               uint
 	  >*);
 }  // namespace fertilized
@@ -67,8 +67,8 @@ namespace serialization {
 
 template <>
 struct version<DifferenceSurfaceCalculator<
-              uint8_t,
-              int16_t,
+              float,
+              float,
               uint
 	  >> {
     typedef mpl::int_<FERTILIZED_VERSION_COUNT> type;
